Better fix for infinite loop in forge3
This fix is closer to the actual problem than the previous fix. The
problem is that a push_trigger is setup in a way that it's origin2
member is NaN. NPCs that walk into this push_trigger are therefore
moved to the position NaN which causes the lerp function to go in an
infinite loop. The previous fix checked for NaN in that lerp function.
This fix checks for NaN in the computation of origin2 instead. I still
do not understand why the height is negative. However it seems like
Raven had similary issues at least at some point:
https://github.com/jedis/jedioutcast/blob/master/code/game/g_trigger.cpp#L755