Allow disabling velocity snapping
Turns out that "SnapVector" on the velocity works against strafe jumping
quite a bit. It simply casts the velocity components to int, which just
truncates the decimals for usual compilers, so it reduces speed in any
case. Especially when already quite fast or for high framerates this is
a major factor.

In the multiplayer (Holomatch) the original code was compiled for the
Quake Virtual Machine. The compiler for that seems to actually do proper
rounding when casting to int, which pretty much avoids this problem.

In Jedi Outcast and Jedi Academy singleplayers the snapping is disabled
completely.

So to allow experimenting a bit this introduces the cvar g_snapVelocity
which is enabled by default to have original behaviour. Cheat protected
of course to avoid accidentally playing with no snapping.