Fix occasional menu softlock
Sometimes when returning to mainmenu, the opening animation would not
start. You then could not do anything to get out, kinda softlocking the
game. Turns out the animation did start, but UI_SetActiveMenu kept being
called for "mainMenu" over and over again, restarting the animation
continuously.

Seems like the engine code repeats calls to UI_SetActiveMenu when
fullscreen is false. But fullscreen for main menu should always be true.
In the softlock case, it was set to false in UI_FrameTop_Graphics,
because seemingly ingameFlag was not reset to false correctly. As main
menu should never be ingame, just making sure to reset ingameFlag to
false whenever the main menu is started now.